+++ src/main/java/org/semver4j/Range.java
@@ -5,6 +5,9 @@ import java.util.Objects;
 import static java.lang.String.format;
 import static java.util.Arrays.stream;
 
+/**
+ * Represents single range item.
+ */
 public class Range {
     private final Semver rangeVersion;
     private final RangeOperator rangeOperator;
@@ -22,10 +25,24 @@ public class Range {
         return rangeVersion;
     }
 
+    /**
+     * Check is range is satisfied by given version.
+     *
+     * @param version version to check
+     * @return {@code true} if range is satisfied by version, {@code false} otherwise
+     * @see #isSatisfiedBy(Semver)
+     */
     public boolean isSatisfiedBy(String version) {
         return isSatisfiedBy(new Semver(version));
     }
 
+    /**
+     * Check is range is satisfied by given version.
+     *
+     * @param version version to check
+     * @return {@code true} if range is satisfied by version, {@code false} otherwise
+     * @see #isSatisfiedBy(String)
+     */
     public boolean isSatisfiedBy(Semver version) {
         switch (rangeOperator) {
             case EQ:
@@ -40,7 +57,7 @@ public class Range {
                 return version.isGreaterThanOrEqualTo(rangeVersion);
         }
 
-        throw new RuntimeException(format("Code error. Unknown RangeOperator: %s", this.rangeOperator));
+        throw new RuntimeException(format("Unknown RangeOperator: %s", rangeOperator));
     }
 
     @Override
@@ -67,27 +84,27 @@ public class Range {
 
     public enum RangeOperator {
         /**
-         * The version and the requirement are equivalent
+         * The version and the requirement are equivalent.
          */
         EQ("="),
 
         /**
-         * The version is lower than the requirement
+         * The version is lower than the requirement.
          */
         LT("<"),
 
         /**
-         * The version is lower than or equivalent to the requirement
+         * The version is lower than or equivalent to the requirement.
          */
         LTE("<="),
 
         /**
-         * The version is greater than the requirement
+         * The version is greater than the requirement.
          */
         GT(">"),
 
         /**
-         * The version is greater than or equivalent to the requirement
+         * The version is greater than or equivalent to the requirement.
          */
         GTE(">="),
         ;
@@ -98,6 +115,11 @@ public class Range {
             this.string = string;
         }
 
+        /**
+         * String representation of the range operator.
+         *
+         * @return range operator as string
+         */
         public String asString() {
             return string;
         }
@@ -109,7 +131,7 @@ public class Range {
             return stream(values())
                     .filter(rangeOperator -> rangeOperator.asString().equals(string))
                     .findFirst()
-                    .orElseThrow(() -> new IllegalArgumentException(format("Range operator for '%s' not found.", string)));
+                    .orElseThrow(() -> new IllegalArgumentException(format("Range operator for '%s' not found", string)));
         }
     }
 }
