+++ src/main/java/org/semver4j/RangesList.java
@@ -4,6 +4,36 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.StringJoiner;
 
+/**
+ * <p>Represents set of single {@link Range}'s.</p>
+ * It's very convenient way to handle complex ranges string.<br>
+ * <br>
+ * Following range:
+ *
+ * <pre>
+ * <= 2.6.8 || >= 3.0.0 <= 3.0.1
+ * </pre>
+ * <p>
+ * is parsed into:
+ *
+ * <pre>
+ * RangesList[
+ *      rangesList=[
+ *          [=2.6.8],
+ *          [=3.0.0, =3.0.1]
+ *      ]
+ * ]
+ * </pre>
+ * <p>
+ * That means, one of ranges {@code [=2.6.8]} or {@code [=3.0.0, =3.0.1]} must by full applied.<br>
+ * So that example allows pass following version:
+ * <ul>
+ *     <li>{@code 2.6.8} - because of range {@code [=2.6.8]}</li>
+ *     <li>{@code 3.0.0} - because of range {@code [=3.0.0, =3.0.1]}</li>
+ *     <li>{@code 3.0.1} - because of range {@code [=3.0.0, =3.0.1]}</li>
+ * </ul>
+ * Any other versions not satisfied this range.
+ */
 public class RangesList {
     private final List<List<Range>> rangesList = new ArrayList<>();
