+++ src/test/java/org/semver4j/SemverTest.java
@@ -832,149 +832,291 @@ class SemverTest {
         assertThat(lowerThan1).isEqualTo(lowerThan2);
     }
 
+    @ParameterizedTest
+    @MethodSource("getParameters")
+    public void shouldCheckSatisfies(String version, String range, boolean expected) {
+        //given
+        Semver semver = new Semver(version);
+
+        //when
+        boolean satisfies = semver.satisfies(range);
+
+        //then
+        assertThat(satisfies).isEqualTo(expected);
+    }
+
+    public static Stream<Arguments> getParameters() {
+        return Stream.of(
+                // Fully-qualified versions:
+                arguments("1.0.0", "1.0.0", true),
+                arguments("1.0.0", "=1.0.0", true),
+                arguments("1.2.3", "1.2.3", true),
+                arguments("1.2.4", "1.2.3", false),
+                arguments("1.0.0-setup-20220428123901", "1.0.0-setup-20220428123901", true),
+
+                // Minor versions:
+                arguments("1.2.3", "1.2", true),
+                arguments("1.2.4", "1.3", false),
+
+                // Major versions:
+                arguments("1.2.3", "1", true),
+                arguments("1.2.4", "2", false),
+
+                // Hyphen ranges:
+                arguments("1.2.4-beta+exp.sha.5114f85", "1.2.3 - 2.3.4", false),
+                arguments("1.2.4", "1.2.3 - 2.3.4", true),
+                arguments("1.2.3", "1.2.3 - 2.3.4", true),
+                arguments("2.3.4", "1.2.3 - 2.3.4", true),
+                arguments("2.3.0-alpha", "1.2.3 - 2.3.0-beta", true),
+                arguments("2.3.4", "1.2.3 - 2.3", true),
+                arguments("2.3.4", "1.2.3 - 2", true),
+                arguments("4.4.0", "3.X - 4.X", true),
+                arguments("1.0.0", "1.2.3 - 2.3.4", false),
+                arguments("3.0.0", "1.2.3 - 2.3.4", false),
+                arguments("2.4.3", "1.2.3 - 2.3", false),
+                arguments("2.3.0-rc1", "1.2.3 - 2.3.0-beta", false),
+                arguments("3.0.0", "1.2.3 - 2", false),
+
+                // Wildcard ranges:
+                arguments("3.1.5", "", true),
+                arguments("3.1.5", "*", true),
+                arguments("0.0.0", "*", true),
+                arguments("1.0.0-beta", "*", false),
+                arguments("3.1.5-beta", "3.1.x", false),
+                arguments("3.1.5-beta+exp.sha.5114f85", "3.1.x", false),
+                arguments("3.1.5+exp.sha.5114f85", "3.1.x", true),
+                arguments("3.1.5", "3.1.x", true),
+                arguments("3.1.5", "3.1.X", true),
+                arguments("3.1.5", "3.x", true),
+                arguments("3.1.5", "3.*", true),
+                arguments("3.1.5", "3.1", true),
+                arguments("3.1.5", "3", true),
+                arguments("3.2.5", "3.1.x", false),
+                arguments("3.0.5", "3.1.x", false),
+                arguments("4.0.0", "3.x", false),
+                arguments("2.0.0", "3.x", false),
+                arguments("3.2.5", "3.1", false),
+                arguments("3.0.5", "3.1", false),
+                arguments("4.0.0", "3", false),
+                arguments("2.0.0", "3", false),
+
+                // Tilde ranges:
+                arguments("1.2.4-beta", "~1.2.3", false),
+                arguments("1.2.4-beta+exp.sha.5114f85", "~1.2.3", false),
+                arguments("1.2.3", "~1.2.3", true),
+                arguments("1.2.7", "~1.2.3", true),
+                arguments("1.2.2", "~1.2", true),
+                arguments("1.2.0", "~1.2", true),
+                arguments("1.3.0", "~1", true),
+                arguments("1.0.0", "~1", true),
+                arguments("1.2.3", "~1.2.3-beta.2", true),
+                arguments("1.2.3-beta.4", "~1.2.3-beta.2", true),
+                arguments("1.2.4", "~1.2.3-beta.2", true),
+                arguments("1.3.0", "~1.2.3", false),
+                arguments("1.2.2", "~1.2.3", false),
+                arguments("1.1.0", "~1.2", false),
+                arguments("1.3.0", "~1.2", false),
+                arguments("2.0.0", "~1", false),
+                arguments("0.0.0", "~1", false),
+                arguments("1.2.3-beta.1", "~1.2.3-beta.2", false),
+                arguments("0.0.7", "~1.9.1-6", false),
+
+                // Caret ranges:
+                arguments("16.14.0", "^16.0.0-0", true),
+                arguments("1.2.3", "^1.2.3", true),
+                arguments("1.2.4", "^1.2.3", true),
+                arguments("1.3.0", "^1.2.3", true),
+                arguments("0.2.3", "^0.2.3", true),
+                arguments("0.2.4", "^0.2.3", true),
+                arguments("0.0.3", "^0.0.3", true),
+                arguments("0.0.3+exp.sha.5114f85", "^0.0.3", true),
+                arguments("0.0.3", "^0.0.3-beta", true),
+                arguments("0.0.3-pr.2", "^0.0.3-beta", true),
+                arguments("1.2.2", "^1.2.3", false),
+                arguments("2.0.0", "^1.2.3", false),
+                arguments("0.2.2", "^0.2.3", false),
+                arguments("0.3.0", "^0.2.3", false),
+                arguments("0.0.4", "^0.0.3", false),
+                arguments("0.0.3-alpha", "^0.0.3-beta", false),
+                arguments("0.0.4", "^0.0.3-beta", false),
+
+                // Comparators:
+                arguments("2.0.0", "=2.0.0", true),
+                arguments("2.0.0", "=2.0", true),
+                arguments("2.0.1", "=2.0", true),
+                arguments("2.0.0", "=2", true),
+                arguments("2.0.1", "=2", true),
+                arguments("2.0.1", "=2.0.0", false),
+                arguments("1.9.9", "=2.0.0", false),
+                arguments("1.9.9", "=2.0", false),
+                arguments("1.9.9", "=2", false),
+
+                arguments("2.0.1", ">2.0.0", true),
+                arguments("3.0.0", ">2.0.0", true),
+                arguments("3.0.0", ">2.0", true),
+                arguments("3.0.0", ">2", true),
+                arguments("2.0.0", ">2.0.0", false),
+                arguments("1.9.9", ">2.0.0", false),
+                arguments("2.0.0", ">2.0", false),
+                arguments("1.9.9", ">2.0", false),
+                arguments("2.0.1", ">2", false),
+                arguments("2.0.0", ">2", false),
+                arguments("1.9.9", ">2", false),
+
+                arguments("1.9.9", "<2.0.0", true),
+                arguments("1.9.9", "<2.0", true),
+                arguments("1.9.9", "<2", true),
+                arguments("2.0.0", "<2.0.0", false),
+                arguments("2.0.1", "<2.0.0", false),
+                arguments("3.0.0", "<2.0.0", false),
+                arguments("2.0.0", "<2.0", false),
+                arguments("2.0.1", "<2.0", false),
+                arguments("3.0.0", "<2.0", false),
+                arguments("2.0.0", "<2", false),
+                arguments("2.0.1", "<2", false),
+                arguments("3.0.0", "<2", false),
+
+                arguments("2.0.0", ">=2.0.0", true),
+                arguments("2.0.1", ">=2.0.0", true),
+                arguments("3.0.0", ">=2.0.0", true),
+                arguments("2.0.0", ">=2.0", true),
+                arguments("3.0.0", ">=2.0", true),
+                arguments("2.0.0", ">=2", true),
+                arguments("2.0.1", ">=2", true),
+                arguments("3.0.0", ">=2", true),
+                arguments("1.9.9", ">=2.0.0", false),
+                arguments("1.9.9", ">=2.0", false),
+                arguments("1.9.9", ">=2", false),
+                arguments("3.3.1-alpha", ">=2.4.x", false),
+
+                arguments("3.3.1", ">=2.4.x", true),
+                arguments("1.9.9", "<=2.0.0", true),
+                arguments("2.0.0", "<=2.0.0", true),
+                arguments("1.9.9", "<=2.0", true),
+                arguments("2.0.0", "<=2.0", true),
+                arguments("2.0.1", "<=2.0", true),
+                arguments("1.9.9", "<=2", true),
+                arguments("2.0.0", "<=2", true),
+                arguments("2.0.1", "<=2", true),
+                arguments("2.0.1", "<=2.0.0", false),
+                arguments("3.0.0", "<=2.0.0", false),
+                arguments("3.0.0", "<=2.0", false),
+                arguments("3.0.0", "<=2", false),
+
+                // AND ranges:
+                arguments("2.0.1", ">2.0.0 <3.0.0", true),
+                arguments("2.0.1", ">2.0 <3.0", false),
+
+                arguments("1.2.0", "1.2 <1.2.8", true),
+                arguments("1.2.7", "1.2 <1.2.8", true),
+                arguments("1.1.9", "1.2 <1.2.8", false),
+                arguments("1.2.9", "1.2 <1.2.8", false),
+
+                // OR ranges:
+                arguments("1.2.3", "1.2.3 || 1.2.4", true),
+                arguments("1.2.4", "1.2.3 || 1.2.4", true),
+                arguments("1.2.5", "1.2.3 || 1.2.4", false),
+
+                // Complex ranges:
+                arguments("1.2.2", ">1.2.1 <1.2.8 || >2.0.0", true),
+                arguments("1.2.7", ">1.2.1 <1.2.8 || >2.0.0", true),
+                arguments("2.0.1", ">1.2.1 <1.2.8 || >2.0.0", true),
+                arguments("1.2.1", ">1.2.1 <1.2.8 || >2.0.0", false),
+                arguments("2.0.0", ">1.2.1 <1.2.8 || >2.0.0", false),
+
+                arguments("1.2.2", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", true),
+                arguments("1.2.7", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", true),
+                arguments("2.0.1", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", true),
+                arguments("2.5.0", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", true),
+                arguments("1.2.1", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", false),
+                arguments("1.2.8", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", false),
+                arguments("2.0.0", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", false),
+                arguments("3.0.0", ">1.2.1 <1.2.8 || >2.0.0 <3.0.0", false),
+
+                arguments("1.2.2", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", true),
+                arguments("1.2.7", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", true),
+                arguments("2.0.1", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", true),
+                arguments("2.5.0", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", true),
+                arguments("1.2.1", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", false),
+                arguments("1.2.8", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", false),
+                arguments("2.0.0", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", false),
+                arguments("3.0.0", "1.2.2 - 1.2.7 || 2.0.1 - 2.9.9", false),
+
+                arguments("1.2.0", "1.2 <1.2.8 || >2.0.0", true),
+                arguments("1.2.7", "1.2 <1.2.8 || >2.0.0", true),
+                arguments("1.2.7", "1.2 <1.2.8 || >2.0.0", true),
+                arguments("2.0.1", "1.2 <1.2.8 || >2.0.0", true),
+                arguments("1.1.0", "1.2 <1.2.8 || >2.0.0", false),
+                arguments("1.2.9", "1.2 <1.2.8 || >2.0.0", false),
+                arguments("2.0.0", "1.2 <1.2.8 || >2.0.0", false),
+                arguments("2.6.9", "<= 2.6.8 || >= 3.0.0 <= 3.0.1", false),
+
+                arguments("1.2.2", " ~> 1.2.3 ", false),
+                arguments("1.2.3", " ~> 1.2.3 ", true),
+                arguments("1.2.4", " ~> 1.2.3 ", true),
+                arguments("1.3.0", " ~> 1.2.3 ", false),
+                arguments("2.2.0", " ~> 2.2 ", true),
+                arguments("2.3.0", " ~> 2.2 ", false),
+
+                arguments("0.0.9", "[1.0,2.0]", false),
+                arguments("1.0.0", "[1.0,2.0]", true),
+                arguments("2.0.0", "[1.0,2.0]", true),
+                arguments("1.5.6", "[1.0,2.0]", true),
+                arguments("2.0.1", "[1.0,2.0]", false),
+
+                arguments("2.0.0", "[1.0,2.0[", false),
+                arguments("1.0.0", "[1.0,2.0[", true),
+                arguments("0.0.9", "[1.0,2.0[", false),
+                arguments("2.0.1", "[1.0,2.0[", false),
+                arguments("1.5.6", "[1.0,2.0[", true),
+
+                arguments("1.0.0", "]1.0,2.0]", false),
+                arguments("1.5.6", "]1.0,2.0]", true),
+                arguments("2.0.0", "]1.0,2.0]", true),
+                arguments("2.0.1", "]1.0,2.0]", false),
+
+                arguments("1.0.0", "]1.0,2.0[", false),
+                arguments("2.0.0", "]1.0,2.0[", false),
+                arguments("1.5.6", "]1.0,2.0[", true),
+
+                arguments("1.0.0", "[1.0,)", true),
+                arguments("1.0.100", "[1.0,)", true),
+                arguments("100.0.1", "[1.0,)", true),
+                arguments("0.0.9", "[1.0,)", false),
+
+                arguments("1.0.0", "]1.0,)", false),
+                arguments("1.0.100", "]1.0,)", true),
+                arguments("100.0.1", "]1.0,)", true),
+                arguments("0.0.9", "]1.0,)", false),
+
+                arguments("2.0.0", "(,2.0]", true),
+                arguments("2.0.10", "(,2.0]", false),
+                arguments("3.0.10", "(,2.0]", false),
+                arguments("1.0.100", "(,2.0]", true),
+                arguments("0.0.9", "(,2.0]", true),
+
+                arguments("2.0.0", "(,2.0[", false),
+                arguments("2.0.10", "(,2.0[", false),
+                arguments("3.0.10", "(,2.0[", false),
+                arguments("1.0.100", "(,2.0[", true),
+                arguments("0.0.9", "(,2.0[", true),
+
+                arguments("1.2.0", "1.2.+", true),
+                arguments("1.1.90", "1.2.+", false),
+                arguments("1.3.0", "1.2.+", false),
+                arguments("1.2.90", "1.2.+", true),
+
+                arguments("1.0.0", "1.+", true),
+                arguments("2.0.0", "1.+", false),
+                arguments("2.0.1", "1.+", false),
+                arguments("1.3.0", "1.+", true),
+                arguments("1.2.90", "1.+", true),
+
+                arguments("0.0.0", "latest.integration", true)
+        );
+    }
+
     private static String repeat(String s, int n) {
         return join("", nCopies(n, s));
     }
-
-    //    @Test
-//    public void statisfies_calls_the_requirement() {
-//        Requirement req = mock(Requirement.class);
-//        Semver semver = new Semver("1.2.2");
-//        semver.satisfies(req);
-//        verify(req).isSatisfiedBy(semver);
-//    }
-//
-//
-//    @Test
-//    public void withIncMinor_test() {
-//        Semver semver = new Semver("1.2.3-Beta.4+SHA123456789");
-//        semver.withIncMinor(2).isEqualTo("1.4.3-Beta.4+SHA123456789");
-//    }
-//
-//    @Test
-//    public void withIncPatch_test() {
-//        Semver semver = new Semver("1.2.3-Beta.4+SHA123456789");
-//        semver.withIncPatch(2).isEqualTo("1.2.5-Beta.4+SHA123456789");
-//    }
-//
-//    @Test
-//    public void withClearedSuffix_test() {
-//        Semver semver = new Semver("1.2.3-Beta.4+SHA123456789");
-//        semver.withClearedSuffix().isEqualTo("1.2.3+SHA123456789");
-//    }
-//
-//    @Test
-//    public void withClearedBuild_test() {
-//        Semver semver = new Semver("1.2.3-Beta.4+sha123456789");
-//        semver.withClearedBuild().isEqualTo("1.2.3-Beta.4");
-//    }
-//
-//    @Test
-//    public void withClearedBuild_test_multiple_hyphen_signs() {
-//        Semver semver = new Semver("1.2.3-Beta.4-test+sha12345-6789");
-//        semver.withClearedBuild().isEqualTo("1.2.3-Beta.4-test");
-//    }
-//
-//    @Test
-//    public void withClearedSuffixAndBuild_test() {
-//        Semver semver = new Semver("1.2.3-Beta.4+SHA123456789");
-//        semver.withClearedSuffixAndBuild().isEqualTo("1.2.3");
-//    }
-//
-//    @Test
-//    public void withSuffix_test_change_suffix() {
-//        Semver semver = new Semver("1.2.3-Alpha.4+SHA123456789");
-//        Semver result = semver.withSuffix("Beta.1");
-//
-//        assertEquals("1.2.3-Beta.1+SHA123456789", result.toString());
-////        assertArrayEquals(new String[]{"Beta", "1"}, result.getPreRelease());
-//    }
-//
-//    @Test
-//    public void withSuffix_test_add_suffix() {
-//        Semver semver = new Semver("1.2.3+SHA123456789");
-//        Semver result = semver.withSuffix("Beta.1");
-//
-//        assertEquals("1.2.3-Beta.1+SHA123456789", result.toString());
-////        assertArrayEquals(new String[]{"Beta", "1"}, result.getPreRelease());
-//    }
-//
-//    @Test
-//    public void withBuild_test_change_build() {
-//        Semver semver = new Semver("1.2.3-Alpha.4+SHA123456789");
-//        Semver result = semver.withBuild("SHA987654321");
-//
-//        assertEquals("1.2.3-Alpha.4+SHA987654321", result.toString());
-//        assertEquals("SHA987654321", result.getBuild());
-//    }
-//
-//    @Test
-//    public void withBuild_test_add_build() {
-//        Semver semver = new Semver("1.2.3-Alpha.4");
-//        Semver result = semver.withBuild("SHA987654321");
-//
-//        assertEquals("1.2.3-Alpha.4+SHA987654321", result.toString());
-//        assertEquals("SHA987654321", result.getBuild());
-//    }
-//
-//    @Test
-//    public void nextMajor_test() {
-//        Semver semver = new Semver("1.2.3-beta.4+sha123456789");
-//        semver.nextMajor().isEqualTo("2.0.0");
-//    }
-//
-//    @Test
-//    public void nextMinor_test() {
-//        Semver semver = new Semver("1.2.3-beta.4+sha123456789");
-//        semver.nextMinor().isEqualTo("1.3.0");
-//    }
-//
-//    @Test
-//    public void nextPatch_test() {
-//        Semver semver = new Semver("1.2.3-beta.4+sha123456789");
-//        semver.nextPatch().isEqualTo("1.2.4");
-//    }
-//
-//    @Test
-//    public void toStrict_test() {
-//        String[][] versionGroups = new String[][]{new String[]{"3.0.0-beta.4+sha123456789", "3.0-beta.4+sha123456789", "3-beta.4+sha123456789"}, new String[]{"3.0.0+sha123456789", "3.0+sha123456789", "3+sha123456789"}, new String[]{"3.0.0-beta.4", "3.0-beta.4", "3-beta.4"}, new String[]{"3.0.0", "3.0", "3"},};
-//
-//        Semver.SemverType[] types = new Semver.SemverType[]{Semver.SemverType.NPM, Semver.SemverType.IVY, Semver.SemverType.LOOSE, Semver.SemverType.COCOAPODS,};
-//
-//        for (String[] versions : versionGroups) {
-//            Semver strict = new Semver(versions[0]);
-//            assertEquals(strict, strict.toStrict());
-//            for (Semver.SemverType type : types) {
-//                for (String version : versions) {
-//                    Semver sem = new Semver(version, type);
-//                    assertEquals(strict, sem.toStrict());
-//                }
-//            }
-//        }
-//    }
-//
-
-//
-//
-//    @Test
-//    public void compareTo_without_path_or_minor() {
-//        assertTrue(new Semver("1.2.3", Semver.SemverType.LOOSE).isGreaterThan("1.2"));
-//        assertTrue(new Semver("1.3", Semver.SemverType.LOOSE).isGreaterThan("1.2.3"));
-//        assertTrue(new Semver("1.2.3", Semver.SemverType.LOOSE).isGreaterThan("1"));
-//        assertTrue(new Semver("2", Semver.SemverType.LOOSE).isGreaterThan("1.2.3"));
-//    }
-//
-//    @Test
-//    public void getValue_returns_the_original_value_trimmed_and_with_the_same_case() {
-//        String version = "  1.2.3-BETA.11+sHa.0nSFGKjkjsdf  ";
-//        Semver semver = new Semver(version);
-//        assertEquals("1.2.3-BETA.11+sHa.0nSFGKjkjsdf", semver.getValue());
-//    }
-//
-//    @Test
-//    public void compareTo_with_buildNumber() {
-//        Semver v3 = new Semver("1.24.1-rc3+903423.234");
-//        Semver v4 = new Semver("1.24.1-rc3+903423.235");
-//        assertEquals(0, v3.compareTo(v4));
-//    }
 }
