+++ src/test/java/org/semver4j/internal/StrictParserTest.java
@@ -4,61 +4,64 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.semver4j.SemverException;
+import org.semver4j.internal.StrictParser.Version;
 
 import java.util.stream.Stream;
 
 import static java.lang.String.format;
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
+import static java.util.Collections.singletonList;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 class StrictParserTest {
-//    @ParameterizedTest
-//    @MethodSource("validStrictSemver")
-void shouldParseValidVersions(String version, StrictParser.Version expected) {
-    //given
-//        StrictSemverParser strictSemverParser = new StrictSemverParser();
-//
-//        //when
-//        ParsedVersion actual = strictSemverParser.parse(version);
-//
-//        //then
-//        assertThat(actual).isEqualTo(expected);
-}
+    @ParameterizedTest
+    @MethodSource("validStrictSemver")
+    void shouldParseValidVersions(String version, Version expected) {
+        //given
+        StrictParser strictParser = new StrictParser();
+
+        //when
+        Version actual = strictParser.parse(version);
+
+        //then
+        assertThat(actual).isEqualTo(expected);
+    }
 
     static Stream<Arguments> validStrictSemver() {
         return Stream.of(
-//                arguments("0.0.4", new ParsedVersion(0, 0, 4)),
-//                arguments("1.2.3", new ParsedVersion(1, 2, 3)),
-//                arguments("10.20.30", new ParsedVersion(10, 20, 30)),
-//                arguments("1.1.2-prerelease+meta", new ParsedVersion(1, 1, 2, "prerelease", "meta")),
-//                arguments("1.1.2+meta", new ParsedVersion(1, 1, 2, null, "meta")),
-//                arguments("1.1.2+meta-valid", new ParsedVersion(1, 1, 2, null, "meta-valid")),
-//                arguments("1.0.0-alpha", new ParsedVersion(1, 0, 0, "alpha", null)),
-//                arguments("1.0.0-beta", new ParsedVersion(1, 0, 0, "beta", null)),
-//                arguments("1.0.0-alpha.beta", new ParsedVersion(1, 0, 0, "alpha.beta", null)),
-//                arguments("1.0.0-alpha.beta.1", new ParsedVersion(1, 0, 0, "alpha.beta.1", null)),
-//                arguments("1.0.0-alpha.1", new ParsedVersion(1, 0, 0, "alpha.1", null)),
-//                arguments("1.0.0-alpha0.valid", new ParsedVersion(1, 0, 0, "alpha0.valid", null)),
-//                arguments("1.0.0-alpha.0valid", new ParsedVersion(1, 0, 0, "alpha.0valid", null)),
-//                arguments("1.0.0-alpha-a.b-c-somethinglong+build.1-aef.1-its-okay", new ParsedVersion(1, 0, 0, "alpha-a.b-c-somethinglong", "build.1-aef.1-its-okay")),
-//                arguments("1.0.0-rc.1+build.1", new ParsedVersion(1, 0, 0, "rc.1", "build.1")),
-//                arguments("2.0.0-rc.1+build.123", new ParsedVersion(2, 0, 0, "rc.1", "build.123")),
-//                arguments("1.2.3-beta", new ParsedVersion(1, 2, 3, "beta", null)),
-//                arguments("10.2.3-DEV-SNAPSHOT", new ParsedVersion(10, 2, 3, "DEV-SNAPSHOT", null)),
-//                arguments("1.2.3-SNAPSHOT-123", new ParsedVersion(1, 2, 3, "SNAPSHOT-123", null)),
-//                arguments("1.0.0", new ParsedVersion(1, 0, 0)),
-//                arguments("2.0.0", new ParsedVersion(2, 0, 0)),
-//                arguments("1.1.7", new ParsedVersion(1, 1, 7)),
-//                arguments("2.0.0+build.1848", new ParsedVersion(2, 0, 0, null, "build.1848")),
-//                arguments("2.0.1-alpha.1227", new ParsedVersion(2, 0, 1, "alpha.1227", null)),
-//                arguments("1.0.0-alpha+beta", new ParsedVersion(1, 0, 0, "alpha", "beta")),
-//                arguments("1.2.3----RC-SNAPSHOT.12.9.1--.12+788", new ParsedVersion(1, 2, 3, "---RC-SNAPSHOT.12.9.1--.12", "788")),
-//                arguments("1.2.3----R-S.12.9.1--.12+meta", new ParsedVersion(1, 2, 3, "---R-S.12.9.1--.12", "meta")),
-//                arguments("1.2.3----RC-SNAPSHOT.12.9.1--.12", new ParsedVersion(1, 2, 3, "---RC-SNAPSHOT.12.9.1--.12", null)),
-//                arguments("1.0.0+0.build.1-rc.10000aaa-kk-0.1", new ParsedVersion(1, 0, 0, null, "0.build.1-rc.10000aaa-kk-0.1")),
-////todo need to be handled
-////                arguments("99999999999999999999999.999999999999999999.99999999999999999", new ParsedVersion(99999999999999999999999, 999999999999999999, 99999999999999999)),
-//                arguments("1.0.0-0A.is.legal", new ParsedVersion(1, 0, 0, "0A.is.legal", null))
+                arguments("0.0.4", new Version(0, 0, 4)),
+                arguments("1.2.3", new Version(1, 2, 3)),
+                arguments("10.20.30", new Version(10, 20, 30)),
+                arguments("1.1.2-prerelease+meta", new Version(1, 1, 2, singletonList("prerelease"), singletonList("meta"))),
+                arguments("1.1.2+meta", new Version(1, 1, 2, emptyList(), singletonList("meta"))),
+                arguments("1.1.2+meta-valid", new Version(1, 1, 2, emptyList(), singletonList("meta-valid"))),
+                arguments("1.0.0-alpha", new Version(1, 0, 0, singletonList("alpha"), emptyList())),
+                arguments("1.0.0-beta", new Version(1, 0, 0, singletonList("beta"), emptyList())),
+                arguments("1.0.0-alpha.beta", new Version(1, 0, 0, asList("alpha.beta".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha.beta.1", new Version(1, 0, 0, asList("alpha.beta.1".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha.1", new Version(1, 0, 0, asList("alpha.1".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha0.valid", new Version(1, 0, 0, asList("alpha0.valid".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha.0valid", new Version(1, 0, 0, asList("alpha.0valid".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha-a.b-c-somethinglong+build.1-aef.1-its-okay", new Version(1, 0, 0, asList("alpha-a.b-c-somethinglong".split("\\.")), asList("build.1-aef.1-its-okay".split("\\.")))),
+                arguments("1.0.0-rc.1+build.1", new Version(1, 0, 0, asList("rc.1".split("\\.")), asList("build.1".split("\\.")))),
+                arguments("2.0.0-rc.1+build.123", new Version(2, 0, 0, asList("rc.1".split("\\.")), asList("build.123".split("\\.")))),
+                arguments("1.2.3-beta", new Version(1, 2, 3, singletonList("beta"), emptyList())),
+                arguments("10.2.3-DEV-SNAPSHOT", new Version(10, 2, 3, singletonList("DEV-SNAPSHOT"), emptyList())),
+                arguments("1.2.3-SNAPSHOT-123", new Version(1, 2, 3, singletonList("SNAPSHOT-123"), emptyList())),
+                arguments("1.0.0", new Version(1, 0, 0)),
+                arguments("2.0.0", new Version(2, 0, 0)),
+                arguments("1.1.7", new Version(1, 1, 7)),
+                arguments("2.0.0+build.1848", new Version(2, 0, 0, emptyList(), asList("build.1848".split("\\.")))),
+                arguments("2.0.1-alpha.1227", new Version(2, 0, 1, asList("alpha.1227".split("\\.")), emptyList())),
+                arguments("1.0.0-alpha+beta", new Version(1, 0, 0, singletonList("alpha"), singletonList("beta"))),
+                arguments("1.2.3----RC-SNAPSHOT.12.9.1--.12+788", new Version(1, 2, 3, asList("---RC-SNAPSHOT.12.9.1--.12".split("\\.")), singletonList("788"))),
+                arguments("1.2.3----R-S.12.9.1--.12+meta", new Version(1, 2, 3, asList("---R-S.12.9.1--.12".split("\\.")), singletonList("meta"))),
+                arguments("1.2.3----RC-SNAPSHOT.12.9.1--.12", new Version(1, 2, 3, asList("---RC-SNAPSHOT.12.9.1--.12".split("\\.")), emptyList())),
+                arguments("1.0.0+0.build.1-rc.10000aaa-kk-0.1", new Version(1, 0, 0, emptyList(), asList("0.build.1-rc.10000aaa-kk-0.1".split("\\.")))),
+                arguments("1.0.0-0A.is.legal", new Version(1, 0, 0, asList("0A.is.legal".split("\\.")), emptyList()))
         );
     }
